requirements.txt
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
confection==0.1.5
cymem==2.0.8
easyocr==1.7.2
eng_spacysentiment==2.3.0
filelock==3.16.1
Flask==3.0.3
fsspec==2024.9.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.35.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
marisa-trie==1.2.0
MarkupSafe==2.1.5
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.10.0.84
packaging==24.1
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.4.0
preshed==3.0.9
pyclipper==1.3.0.post5
pydantic==1.10.18
pytesseract==0.3.13
python-bidi==0.6.0
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
rfc3986==1.5.0
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.6
smart-open==6.4.0
sniffio==1.3.1
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.3
thinc==8.1.12
tifffile==2023.7.10
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
triton==3.0.0
typer==0.9.4
typing_extensions==4.12.2
urllib3==2.2.3
wasabi==1.1.3
Werkzeug==3.0.4
zipp==3.20.2
